fix: Rename new class to use io.meeds package - Meeds-io/MIPs#170
