+++ component/api/src/main/java/io/meeds/social/space/plugin/SpaceCategoryLifeCycleEvent.java
@@ -1,0 +1,36 @@
+/**
+ * This file is part of the Meeds project (https://meeds.io/).
+ *
+ * Copyright (C) 2020 - 2024 Meeds Association contact@meeds.io
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
+package io.meeds.social.space.plugin;
+
+import org.exoplatform.social.core.space.model.Space;
+import org.exoplatform.social.core.space.spi.SpaceLifeCycleEvent;
+
+import lombok.Getter;
+
+public class SpaceCategoryLifeCycleEvent extends SpaceLifeCycleEvent {
+
+  @Getter
+  private long categoryId;
+
+  public SpaceCategoryLifeCycleEvent(Space space, String target, long categoryId, Type eventType) {
+    super(space, target, eventType);
+    this.categoryId = categoryId;
+  }
+
+}
