--- component/api/src/main/java/org/exoplatform/social/core/space/spi/SpaceCategoryLifeCycleEvent.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2003-2010 eXo Platform SAS.
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Affero General Public License
- * as published by the Free Software Foundation; either version 3
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, see<http://www.gnu.org/licenses/>.
- */
-package org.exoplatform.social.core.space.spi;
-
-import org.exoplatform.social.core.space.model.Space;
-
-import lombok.Getter;
-
-public class SpaceCategoryLifeCycleEvent extends SpaceLifeCycleEvent {
-
-  @Getter
-  private long categoryId;
-
-  public SpaceCategoryLifeCycleEvent(Space space, String target, long categoryId, Type eventType) {
-    super(space, target, eventType);
-    this.categoryId = categoryId;
-  }
-
-}
